feat: do no exit with error when config file is not set
If the default config file (config.yml) doesn't exist, don't exit with
an error. Instead, use a default configuration file. This removes a
small point of friction when gettting started with go-librespot.

On the other hand, if the config path is set explicitly using
`-config_path`, do exit with an error when it cannot be found (because
presumably the user made a mistake there).