fix: default API to localhost
The previous default was to bind to 0.0.0.0 and allow connections from
anywhere. I think this is an insecure default, the default should be
restricted.

To restore the previous behavior, use:

```yaml
server:
  address: ""
```

For example, I'm using the API from Home Assistant running on the same
Raspberry Pi. There is no reason for the API to be exposed on my local
network. And generally it's a good idea to have safe default values (and
require users to opt-in to less safe options).