feat: rewrite ALSA buffering
This is a partial rewrite of the ALSA output:

  * Reading data from the source and writing data out to the ALSA device
    is now done in a single goroutine. There's no internal ringbuffer
    anymore.
  * Instead of closing and reopening the ALSA device, it now uses
    snd_pcm_drop and snd_pcm_prepare.

Thie main benefit is that when you pause, seek, and then resume the
audio, the audio of the previous playback position is no longer played
for a short bit. Even when pausing or seeking, the buffer wasn't getting
cleared. This was not super noticeable due to the relatively short
buffer size (80ms on my laptop, 160ms on my Raspberry Pi 3) but with the
previous commit increasing the period size this became _very_ noticeable
and grew to 2 seconds.